add tables

+++ PADC-FP-HealthCareDirectory/app/src/main/java/com/padc/healthcaredirectory/data/vos/FaxVO.java
@@ -17,7 +17,7 @@ public class FaxVO {
         return faxId;
     }
 
-    public void setFaxId(int faxId) {
+    public void setFaxId(long faxId) {
         this.faxId = faxId;
     }
